chore(suite): move logic from TransportBackends.tsx to useBridgeDesktopApi hook to make it reusable
